[refactor] 무기 넘버링 수정     - EnemyCharacter : bot타입에 따른 바인딩 무기 ID수정     - MainCharacter : ID수정 중 발견된 에임관련 불필요 노드 제거     - Item/Blueprint/Projectile :         - ProjectileItemBase : int32로 수정에 따른 오류 수정         - ProjectileItem : 필드생성 아이템 ID수정     - Item/Blueprint/Item_Weapon :         - WeaponItemBase : int32로 수정에 따른 오류 수정 및 ID수정         - 필드 생성 무기 ID수정     - AssetManager : 불필요 ID관련 컴포넌트 제거     - Inventory : ID관련 수정     - Weapon :         - WeaponInventory ID관련 수정         - WeaponProjectile , Weapon관련 ID수정, 추가     - 데이터 테이블 추가및 수정 무기에 따라 수정     - Native : UInt8->Int32로 수정

+++ Source/BackStreet/Global/public/BackStreetGameModeBase.h
@@ -66,7 +66,7 @@ public:
 		void PlayCameraShakeEffect(ECameraShakeType EffectType, FVector Location, float Radius = 100.0f);
 
 	UFUNCTION()
-		AItemBase* SpawnItemToWorld(uint8 ItemType, uint8 ItemID, FVector SpawnLocation);
+		AItemBase* SpawnItemToWorld(uint8 ItemType, int32 ItemID, FVector SpawnLocation);
 
 	UFUNCTION()
 		void UpdateCharacterStat(class ACharacterBase* TargetCharacter, FCharacterStatStruct NewStat);
@@ -78,13 +78,13 @@ public:
 		void UpdateWeaponStat(class AWeaponBase* TargetWeapon, FWeaponStatStruct NewStat);
 
 	UFUNCTION()
-		void UpdateWeaponStatWithID(class AWeaponBase* TargetWeapon, const uint8 WeaponID);
+		void UpdateWeaponStatWithID(class AWeaponBase* TargetWeapon, const int32 WeaponID);
 
 	UFUNCTION()
-		void UpdateProjectileStatWithID(class AProjectileBase* TargetProjectile, const uint8 ProjectileID);
+		void UpdateProjectileStatWithID(class AProjectileBase* TargetProjectile, const int32 ProjectileID);
 
 	UFUNCTION(BlueprintCallable, BlueprintPure)
-		FWeaponStatStruct GetWeaponStatInfoWithID(const uint8 WeaponID);
+		FWeaponStatStruct GetWeaponStatInfoWithID(const int32 WeaponID);
 
 	UFUNCTION()
 		FStageEnemyTypeStruct GetStageTypeInfoWithRow(uint16 rowName);
